+++ Trio/Sources/APS/OpenAPSSwift/Models/ForecastResult.swift
@@ -1,6 +1,8 @@
+import Foundation
+
 struct ForecastResult {
-    public let iob: [Double]
-    public let cob: [Double]
-    public let uam: [Double]
-    public let zt: [Double]
+    public let iob: [Decimal]
+    public let cob: [Decimal]
+    public let uam: [Decimal]
+    public let zt: [Decimal]
 }
